crawler-service/requirements.txt
# Web Collector - Python Crawler Service
# Python 3.10+ required

# Web crawling framework (LLM-first async crawler)
crawl4ai>=0.8.0

# Web framework for REST API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# HTTP client for async requests
httpx>=0.26.0

# Data validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# HTML parsing for metadata extraction
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Environment variables
python-dotenv>=1.0.0

# ===== 独立模式依赖（STANDALONE=true 时需要）=====
# SQLite async driver
aiosqlite>=0.19.0
